refactor(examples): extract input handling to separate module
- Create input.rs module with picker and capture control functions
- Extract format_picked_source, open_picker, open_picker_for_stream
- Extract start_capture and stop_capture functions
- Reduce main.rs from 1001 to 835 lines
- No functionality changes, all features preserved
